requirements.txt
beautifulsoup4>=4.12.0
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
numpy>=1.24.0,<2.0.0
openai>=1.3.0
openai-whisper>=20230124
openwakeword>=0.6.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.1
pytest>=8.0.0
requests>=2.31.0
simpleaudio>=1.0.4
sounddevice>=0.4.6
soundfile>=0.12.0
TTS>=0.18.0
torch>=2.2.0
transformers>=4.38.0
tqdm>=4.66.0
